Carry over Joi and callback functions
1. Created a "schemas" folder and transfer Joi to inside her.
2. Created a "controllers" folder and transfer callback functions to inside her.
