+++ routes/api/contacts.js
@@ -1,85 +1,16 @@
 const express = require('express');
-const Joi = require('joi');
 
-const contacts = require('../../models/contacts');
-const { HttpError } = require('../../helpers');
+const controllers = require('../../controllers/controllers');
 const router = express.Router();
 
-const addSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-});
+router.get('/', controllers.getAllContacts);
 
-router.get('/', async (req, res, next) => {
-  try {
-    res.json(await contacts.listContacts());
-  }
-  catch (error) {
-    next(error);
-  };
-});
+router.get('/:contactId', controllers.getById);
 
-router.get('/:contactId', async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contacts.getContactById(contactId);
-    if (!result) {
-      throw HttpError(404, 'Not found');
-    }
-    res.json(result);
-  }
-  catch (error) {
-    next(error);
-  };
-});
+router.post('/', controllers.addNewContact);
 
-router.post('/', async (req, res, next) => {
-  try {
-    const { error } = addSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, `missing ${error.message} field`);
-    }
-    const result = await contacts.addContact(req.body);
-    res.status(201).json(result);
-  }
-  catch(error) {
-    next(error);
-  };
-});
+router.delete('/:contactId', controllers.deleteContact);
 
-router.delete('/:contactId', async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contacts.removeContact(contactId);
-    if (!result) {
-      throw HttpError(404, 'Not found');
-    }
-    res.json({
-      message: 'contact deleted'
-    });
-  }
-  catch(error) {
-    next(error);
-  };
-});
-
-router.put('/:contactId', async (req, res, next) => {
-  try {
-    const { error } = addSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, "missing fields");
-    }
-    const { contactId } = req.params;
-    const result = await contacts.updateContact(contactId, req.body);
-    if (!result) {
-      throw HttpError(404, 'Not found');
-    }
-    res.json(result);
-  }
-  catch (error) {
-    next(error);
-  };
-});
+router.put('/:contactId', controllers.changeContact);
 
 module.exports = router;
