+++ schemas/index.js
@@ -1,0 +1,3 @@
+const schemas = require('./schemas');
+
+module.exports = schemas;
